Change Expressions implementation. Allow $frame and $value bounding.

+++ src/core/Expressions/expression.h
@@ -1,3 +1,19 @@
+// enve - 2D animations software
+// Copyright (C) 2016-2020 Maurycy Liebner
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 #ifndef EXPRESSION_H
 #define EXPRESSION_H
 
@@ -34,7 +50,7 @@ public:
 
     static ResultTester sQrealAnimatorTester;
 
-    bool setAbsFrame(const qreal absFrame);
+    bool setAbsFrame(const int absFrame);
 
     bool isValid();
     bool dependsOn(const Property* const prop);
@@ -42,7 +58,7 @@ public:
     QJSValue evaluate();
     QJSValue evaluate(const qreal relFrame);
 
-    FrameRange identicalRange(const qreal absFrame);
+    FrameRange identicalRange(const int absFrame);
 
     QString bindingsString() const;
     const QString& definitionsString() const { return mDefinitionsStr; }
